Refactor list command into separate classes
